Fix applying code action so it works with an inline edit
Some language servers provide a code action with an edit included in
it, rather than a command.  This can be handled cleanly by
`eglot-execute`.

So update the action to not require a `:command` argument unless
`eglot-execute` is not bound, and to pass the entire
`matching-code-action` to `eglot-execute`, as it can handle both
cases.